cand-/match-counting reworked to rmv use of globals
Memo to future self:

I give up. There's an optimization here that yields 3-4%, but it's unsafe.

When '$pat = qr,$stuff,o', 'm,$pat,o' is faster than 'm,$pat,' in the case
where $pat is created 1x and the m operation, performed inside sub
search_files, is performed a grand total of 90K times over 4 calls to
search_files.

'qr,$stuff,o' should in my understanding create a fully compiled regex,
which use of 'o' in 'm,$pat,o' should not be able to further optimize.

My use of 'm,$pat,o' is apparently unsafe in the general sense because
search_files _could_ be called with different $pat values, with the o flag in
'm,$pat,o' causing all but the first $pat value passed to be ignored (the
first-passed $pat would be used in the later calls) which would result in
defective program behavior.  In the current implementation the same instance
(value) of $pat is passed to each of the 4 calls to search_files, thus
resultant program behavior is not defective purely by coincidence.